3971 refactor joinlogicaloperator inferschema (#3979)
* [#3971] Started

* [#3971] Almost done

* [#3971] Have to rewrite WindowType::equal

* [#3971] Added missing comparison function and cleaned up the old ones@

* [#3971] Should be done

* [#3971] Integrated feedback from Philipp

* [#3971] Fixed TypeInferencePhasetest

* [#3971] Integrated feedback from Ankit

* [#3971] Integrated feedback from Ankit

* [#3971] Integrated feedback from Ankit

* [#3971] Integrated feedback from Ankit

---------

+++ nes-core/tests/UnitTests/Optimizer/Phases/TypeInferencePhaseTest.cpp
@@ -102,7 +102,7 @@ TEST_F(TypeInferencePhaseTest, inferQueryPlan) {
     auto mappedSchema = Schema::create();
     mappedSchema->addField("default_logical$f1", BasicType::INT32);
     mappedSchema->addField("default_logical$f2", BasicType::INT8);
-    mappedSchema->addField("default_logical$f3", BasicType::INT8);
+    mappedSchema->addField("default_logical$f3", BasicType::INT32);
 
     NES_DEBUG("first={} second={}", map->getOutputSchema()->toString(), mappedSchema->toString());
     EXPECT_TRUE(map->getOutputSchema()->equals(mappedSchema));
